Create Pattern_Save.ino
My project is basically where I can change the speed of the LEDs lights to my will. Plus, I can also save it to my laptop, allowing me to keep it forever. This means that if I unplug the arduino, and then I plug it in again, the pattern will be saved and will continue.